feat: adds availability view in professionaldto

+++ src/main/java/com/eng/homecare/request/ProfessionalRequestDTO.java
@@ -14,6 +14,6 @@ public record ProfessionalRequestDTO(
         List<CredentialsRequestDTO> credentials,
         String bio,
         String description,
-        int rate
+        Float rate
 ){
 }
